Use qulonglong instead of int to allow big partitions sizes to be detected properly Backport to 4.2branch of SVN commit 919161 by darioandres
BUG: 1335


svn path=/branches/KDE/4.2/kdeutils/kdf/; revision=919184

+++ disks.cpp
@@ -310,13 +310,13 @@ void DiskEntry::setMounted(bool nowMounted)
   emit mountedChanged();
 }
 
-void DiskEntry::setKBSize(int kb_size)
+void DiskEntry::setKBSize(qulonglong kb_size)
 {
   size=kb_size;
   emit kBSizeChanged();
 }
 
-void DiskEntry::setKBUsed(int kb_used)
+void DiskEntry::setKBUsed(qulonglong kb_used)
 {
   used=kb_used;
   if ( size < (used+avail) ) {  //adjust kBAvail
@@ -326,7 +326,7 @@ void DiskEntry::setKBUsed(int kb_used)
   emit kBUsedChanged();
 }
 
-void DiskEntry::setKBAvail(int kb_avail)
+void DiskEntry::setKBAvail(qulonglong kb_avail)
 {
   avail=kb_avail;
   if ( size < (used+avail) ) {  //adjust kBUsed
